refactor: split trade cargo header and technology slot builder parts (#3878)
Extract Market-tab cargo indicator header, row stepper chrome, and
TechnologyPanel research-slot builder into library parts to keep host
files under the repo size target without changing public APIs.

Refs #3878
