+++ app/lib/features/game/screens/trade/trade_screen_market_row.dart
@@ -294,54 +294,3 @@ class _MarketCommodityRowControls extends StatelessWidget {
     );
   }
 }
-
-/// Compact stepper button used by [_MarketCommodityRowControls]. Uses
-/// an `InkWell` so the chrome stays in the editorial-monocle palette
-/// (no Material elevated buttons or accent splash colours).
-class _StepperButton extends StatelessWidget {
-  const _StepperButton({
-    required this.buttonKey,
-    required this.glyph,
-    required this.semanticLabel,
-    required this.onPressed,
-  });
-
-  final Key buttonKey;
-  final String glyph;
-  final String semanticLabel;
-  final VoidCallback? onPressed;
-
-  @override
-  Widget build(BuildContext context) {
-    final ThemeData theme = Theme.of(context);
-    final TextStyle glyphStyle =
-        (theme.textTheme.titleSmall ?? const TextStyle(fontSize: 14))
-            .copyWith(
-              color: onPressed == null
-                  ? EditorialMonoclePalette.muted
-                  : EditorialMonoclePalette.accent,
-            );
-    final Color borderColor = onPressed == null
-        ? EditorialMonoclePalette.muted
-        : EditorialMonoclePalette.accent;
-    return Semantics(
-      button: true,
-      enabled: onPressed != null,
-      label: semanticLabel,
-      child: InkWell(
-        key: buttonKey,
-        onTap: onPressed,
-        child: Container(
-          width: 28,
-          height: 24,
-          alignment: Alignment.center,
-          decoration: BoxDecoration(
-            color: EditorialMonoclePalette.surface.withValues(alpha: 0.5),
-            border: Border.all(color: borderColor, width: 1),
-          ),
-          child: Text(glyph, style: glyphStyle),
-        ),
-      ),
-    );
-  }
-}
